Rename, add main and docker-compose file
