+++ tests/error/docker-compose.yml
@@ -1,0 +1,50 @@
+volumes:
+  config_volume:
+
+services:
+   amqp:
+      image: rabbitmq:alpine
+
+   node_config:
+    build:
+      context: ../../node_config
+    volumes:
+      -  config_volume:/koinos
+
+   jsonrpc:
+      image: koinos/koinos-jsonrpc:${JSONRPC_TAG:-latest}
+      ports:
+         - "8080:8080"
+      volumes:
+         - config_volume:/root/.koinos
+      depends_on:
+         - node_config
+      command: -a amqp://guest:guest@amqp:5672/ -l /tcp/8080 --log-level debug
+
+   mempool:
+      image: koinos/koinos-mempool:${MEMPOOL_TAG:-latest}
+      command: -a amqp://guest:guest@amqp:5672/
+
+   block_store:
+      image: koinos/koinos-block-store:${BLOCK_STORE_TAG:-latest}
+      command: -a amqp://guest:guest@amqp:5672/
+
+   chain:
+      image: koinos/koinos-chain:${CHAIN_TAG:-latest}
+      volumes:
+         - config_volume:/root/.koinos
+      depends_on:
+         - node_config
+      command: -a amqp://guest:guest@amqp:5672/
+
+   p2p:
+      image: koinos/koinos-p2p:${P2P_TAG:-latest}
+      command: -a amqp://guest:guest@amqp:5672/ -s bucket_brigade
+
+   block_producer:
+      image: koinos/koinos-block-producer:${BLOCK_PRODUCER_TAG:-latest}
+      volumes:
+         - config_volume:/root/.koinos
+      depends_on:
+         - node_config
+      command: -a amqp://guest:guest@amqp:5672/ --algorithm=federated --gossip-production=false
